% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMoves.R
\name{plotMoves}
\alias{plotMoves}
\title{two-dimensional Brownian motion plot}
\usage{
plotMoves(moves)
}
\arguments{
\item{moves}{array of moves}
}
\description{
Plot two-dimensional Brownian motion
}
\examples{
moves <- genMoves()
plotMoves(moves)
}

