% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMoves.R
\name{genMoves}
\alias{genMoves}
\title{two-dimensional Brownian motion}
\usage{
genMoves(N = 1000, sigma = 0.1)
}
\arguments{
\item{N}{number of moves}

\item{sigma}{SD of moves}
}
\description{
simulate two-dimensional Brownian motion
}
\examples{
moves <- genMoves()
plotMoves(moves)
generate N values with mean of 0 and standard deviation of 1
generate N values with mean of 0 and standard deviation of sigma^2
added epsilon value to the first entry
update moves vector to be moves[i-1]*epsilon[i]
combine two vectors into array for output
output array
}

